+++ Engine/Project.h
@@ -246,7 +246,7 @@ public:
 
     bool isFrameRangeLocked() const;
     
-    void getFrameRange(int* first,int* last) const;
+    void getFrameRange(double* first,double* last) const;
     
     void unionFrameRangeWith(int first,int last);
